Removed MCU_UART header; Fixed CS2
CS2 was shorted to GND. This has been fixed by removing the UART header and wiring CS2 to one of the two now free pins.
See: R0.0 for more information on this issue.
From now on, please make use of USB to control the Pico.

As part of this fix, the traces and vias around the 5v trace to the header have been updated to be more direct.